created rentals list for customer
